backend/requirements.txt
fastapi==0.128.3
uvicorn==0.40.0
sqlalchemy==2.0.46
alembic==1.14.0
psycopg2-binary>=2.9.9
pydantic==2.12.5
pydantic-settings==2.12.0
python-jose==3.5.0
passlib[argon2]==1.7.4
argon2-cffi>=23.1.0
bcrypt==5.0.0
python-dotenv==1.2.1
email-validator==2.1.0
python-multipart>=0.0.6
pdfplumber>=0.11.0
reportlab>=4.0.0
langgraph>=0.2.0
langchain-openai>=0.2.0
openai>=1.0.0
langchain-core>=0.3.0
boto3>=1.35.0
playwright>=1.40.0
jinja2>=3.1.0
weasyprint>=62.0
razorpay>=1.4.0
redis>=5.0.0
pytest>=7.0.0
httpx>=0.25.0
pytest-asyncio>=0.21.0